제네릭 클래스 예제 및 io 정리

+++ lecture/02_java/chap11_io_lecture_source/src/main/java/com/ohgiraffers/section03/filterstream/Application4.java
@@ -14,15 +14,27 @@ public class Application4 {
         memArr[3] = new MemberDTO("user04", "pass04", "일천", "hong123@email.com", 20, '여');
 
         ObjectOutputStream oos = null;
-        /*
-        try {
-            oos = new ObjectOutputStream(
-                    new FileOutputStream(
-                            "src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt", true
-                    )
-            );
 
-            memArr.length로 넣을 시 null값도 객체가 들어간건가?
+        File ObjFile = new File("src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt");
+
+        // 최초 생성시에는 헤더가 생성되도록 하고 이후 추가할 때는 추가적으로 헤더가 생성되지 않도록 재정의한 스트림 사용
+
+        try {
+            if(!ObjFile.exists()){
+                oos = new ObjectOutputStream( // 최초 선언시 헤더 생성
+                        new FileOutputStream(
+                                "src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt", true
+                        )
+                );
+            } else {
+                oos = new MyOutput( // 추가 생성 시 헤더 X
+                        new FileOutputStream(
+                                "src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt", true
+                        )
+                );
+            }
+
+            // 온전한 객체만 들어가도록 실체가 있는 수만큼 반복
             for (int i = 0; i < 4; i++) {
                 oos.writeObject(memArr[i]);
             }
@@ -35,9 +47,30 @@ public class Application4 {
                 throw new RuntimeException(e);
             }
         }
-        */
-//
 
+// 테스트 흔적
+/*
+        try {
+            oos = new ObjectOutputStream(
+                    new FileOutputStream(
+                            "src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt", true
+                    )
+            );
+
+            // memArr.length로 넣을 시 null값도 객체가 들어간건가?
+            // 객체가 아닌 무언가가 들어가지 않도록 본인이 의도한 값만 들어가도록 잘 설정하고 고려해주어야 한다.
+            for (int i = 0; i < 4; i++) {
+                oos.writeObject(memArr[i]);
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        } finally {
+            try {
+                if (oos != null) oos.close();
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        }
 
         try {
             oos = new MyOutput(
@@ -58,7 +91,10 @@ public class Application4 {
             }
         }
 
-  //
+ */
+
+        // null 값도 insert할 때 들어가서 추가로 insert하고 읽으려고 하면 받아오는 배열 크기가 더 커야 했다.
+        // ex. memArr[10]을 3번 넣었을 때, newMemArr의 크기는 newMemArr[30]이 되었어야 했다. memArr.length로 insert하고 memArr.length로 받아와 크기가 부족하여 에러가 발생했다.
         MemberDTO[] newMemArr = new MemberDTO[memArr.length];
 
         // 객체 배열에 입력을 넣어올 때, 이미 기존의 배열에 가지고 있는 헤더와 나중에 추가되는 객체의 헤더가 충돌하는 일 발생
